src/requirements/test.txt
-e git+https://github.com/robertavram/django-storages.git@1.6.5.3#egg=django-storages
-e git+https://github.com/unicef/djangosaml2.git@v0.16.11.2#egg=djangosaml2
amqp==2.2.2
appnope==0.1.0            # via ipython
asn1crypto==0.24.0
attrs==18.1.0             # via pytest
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage==0.20.2
babel==2.5.3
backcall==0.1.0           # via ipython
billiard==3.5.0.3
carto==1.2.2
celery==4.1.0
cffi==1.11.4
cookies==2.2.1            # via responses
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.1.4
decorator==4.2.1
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.5
dj-static==0.0.6
django-analytical==2.2.2
django-appconf==1.0.2
django-autocomplete-light==3.2.10
django-celery-beat==1.1.1
django-celery-email==2.0
django-celery-results==1.0.1
django-contrib-comments==1.8.0
django-cors-headers==2.2
django-debug-toolbar==1.9.1
django-easy-pdf==0.1.1
django-filter==1.0.2
django-fsm==2.4.0
django-generic-links==0.4.0
django-geojson==2.11
django-hstore==1.4.2
django-import-export==1.0
django-js-asset==1.1.0
django-leaflet==0.23
django-logentry-admin==1.0.3
django-model-utils==2.6
django-mptt==0.9
django-ordered-model==1.4.3
django-post-office==3.0.4
django-redis-cache==1.7.1
django-rest-swagger==2.1.0
django-tenant-schemas==1.9.0
django-waffle==0.14
django==1.10.8
djangorestframework-csv==2.1.0
djangorestframework-hstore==1.3
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
djangorestframework-xml==1.3
djangorestframework==3.5.3
djrill==2.1.0
drf-nested-routers==0.90.2
drfpasswordless==1.1.5
et-xmlfile==1.0.1
factory-boy==2.8
faker==0.8.12
fancycompleter==0.8       # via pdbpp
flake8-html==0.4.0
flake8==3.5.0
flower==0.9.2
freezegun==0.3.10
future==0.15.2
gdal==1.10.0
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
ipython-genutils==0.2.0   # via traitlets
ipython==6.3.1
isort==4.3.4
itypes==1.1.0
jdcal==1.3
jedi==0.12.0              # via ipython
jinja2==2.10
jsonfield==2.0.2
kombu==4.1.0
markupsafe==1.0
mccabe==0.6.1             # via flake8
mock==2.0.0
more-itertools==4.1.0     # via pytest
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.0
parso==0.2.0              # via jedi
paste==2.0.3
pbr==3.1.1                # via mock
pdbpp==0.9.2
pexpect==4.5.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.0.0
pluggy==0.6.0             # via pytest, tox
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.7.4
ptyprocess==0.5.2         # via pexpect
py==1.5.3                 # via pytest, tox
pycodestyle==2.3.1        # via flake8
pycparser==2.18
pycryptodomex==3.4.12
pyflakes==1.6.0           # via flake8
pygments==2.2.0           # via flake8-html, ipython, pdbpp
pyjwt==1.5.3
pyopenssl==17.5.0
pypdf2==1.26.0
pyrestcli==0.6.4
pysaml2==4.4.0
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-echo==1.6.0
pytest-html==1.17.0
pytest-metadata==1.7.0    # via pytest-html
pytest-pythonpath==0.7.2
pytest==3.5.1             # via pytest-cov, pytest-django, pytest-echo, pytest-html, pytest-metadata, pytest-pythonpath
python-dateutil==2.5.3
pytz==2018.3
pyyaml==3.12
raven==6.2.1
redis==2.10.6
reportlab==3.3.0
repoze.who==2.3
requests==2.11.1
responses==0.8.1
simplegeneric==0.8.1      # via ipython
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
tablib==0.12.1
tenant-schemas-celery==0.1.7
text-unidecode==1.2       # via faker
tornado==4.2.0
tox==3.0.0
traitlets==4.3.2          # via ipython
unicodecsv==0.14.1
uritemplate==3.0.0
vine==1.1.4
virtualenv==15.2.0        # via tox
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1
webob==1.7.4
wmctrl==0.3               # via pdbpp
xhtml2pdf==0.2.1
xlrd==1.0.0
xlwt==1.3.0
zope.interface==4.4.3
